requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning & NLP
torch>=1.9.0
transformers>=4.18.0
nltk>=3.6.5
spacy>=3.2.0
finbert==0.1.0
textblob>=0.15.3
sentence-transformers>=2.2.0
pytorch-lightning>=1.6.0

# Web Scraping
beautifulsoup4>=4.9.3
selenium>=4.1.0
requests>=2.26.0
tweepy>=4.10.0
newspaper3k>=0.2.8
scrapy>=2.5.0

# Data Processing & Streaming
apache-kafka-python>=2.0.2
pyyaml>=6.0
python-dotenv>=0.19.0
pymongo>=3.12.0

# Visualization & Dashboard
plotly>=5.5.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
streamlit>=1.0.0

# Testing & Quality
pytest>=7.0.0
pytest-cov>=3.0.0
great-expectations>=0.15.0

# Development Tools
black>=22.3.0
flake8>=4.0.1
isort>=5.10.0
jupyter>=1.0.0
notebook>=6.4.0
pre-commit>=2.15.0

# Database & Storage
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.0.0

# API & Services
fastapi>=0.75.0
uvicorn>=0.17.0
newsapi-python>=0.2.6
python-multipart>=0.0.5

# Monitoring & MLOps
mlflow>=1.26.0
prometheus-client>=0.14.0
evidently>=0.1.0
psutil>=5.9.0
